Kelsey Version 12.8
I did what you said about the moving things before setup, but the text was still behind. So I made a PFont variable thats global and put the font code in the draw function to repeat. Came out with a cool animation, but is VERY VERY SLOW. Also the "c key" part of the output code does not work